+++ webapp/tests/client/ccbilling-pdf-service.test.js
@@ -11,7 +11,7 @@ vi.mock('../../src/lib/client/pdf-utils.js', () => ({
 }));
 
 // Mock the ParserFactory module
-vi.mock('../src/lib/ccbilling-parsers/parser-factory.js', () => ({
+vi.mock('../../src/lib/utils/ccbilling-parsers/parser-factory.js', () => ({
 	ParserFactory: vi.fn()
 }));
 
@@ -25,7 +25,7 @@ describe('PDFService', () => {
 
 		// Get the mocked modules
 		const pdfUtilsModule = await import('../src/lib/utils/pdf-utils.js');
-		const parserFactoryModule = await import('../src/lib/ccbilling-parsers/parser-factory.js');
+		const parserFactoryModule = await import('../../src/lib/utils/ccbilling-parsers/parser-factory.js');
 		
 		mockPDFUtils = pdfUtilsModule.PDFUtils;
 		mockParserFactory = parserFactoryModule.ParserFactory;
